+++ src/main/java/de/samply/converter/Converter.java
@@ -1,10 +1,10 @@
 package de.samply.converter;
 
-import de.samply.template.conversion.ConversionTemplate;
+import de.samply.template.ConverterTemplate;
 import reactor.core.publisher.Flux;
 
 public interface Converter<I,O> {
-  Flux<O> convert (Flux<I> input, ConversionTemplate template);
+  Flux<O> convert (Flux<I> input, ConverterTemplate template);
   Format getInputFormat();
   Format getOutputFormat();
 }