+++ src/main/java/de/samply/converter/ConverterGroup.java
@@ -1,6 +1,6 @@
 package de.samply.converter;
 
-import de.samply.template.conversion.ConversionTemplate;
+import de.samply.template.ConverterTemplate;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
 import reactor.core.publisher.Flux;
@@ -14,7 +14,7 @@ public class ConverterGroup<I, O> implements Converter<I, O> {
   }
 
   @Override
-  public Flux<O> convert(Flux<I> input, ConversionTemplate template) {
+  public Flux<O> convert(Flux<I> input, ConverterTemplate template) {
     AtomicReference<Flux> tempFlux = new AtomicReference<>(input);
     converters.forEach(converter -> tempFlux.set(converter.convert(tempFlux.get(), template)));
     return tempFlux.get();