+++ src/test/java/de/samply/template/ConverterTemplateManagerTest.java
@@ -1,19 +1,19 @@
-package de.samply.container.template;
+package de.samply.template;
 
-import de.samply.template.conversion.ConversionTemplateManager;
+import de.samply.template.ConverterTemplateManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 @Disabled
-class ConversionTemplateManagerTest {
+class ConverterTemplateManagerTest {
 
-  private ConversionTemplateManager templates;
+  private ConverterTemplateManager templates;
   private String templatesDirectory = "./templates";
 
   @BeforeEach
   void setUp() {
-    this.templates = new ConversionTemplateManager(templatesDirectory);
+    this.templates = new ConverterTemplateManager(templatesDirectory);
     //TODO
   }